+++ addons/mail/tests/test_mail_tools.py
@@ -3,7 +3,6 @@
 
 from odoo.addons.mail.tests.common import MailCommon
 from odoo.tests import tagged, users
-from odoo import tools
 
 
 @tagged('mail_tools', 'res_partner')
@@ -21,30 +20,6 @@ class TestMailTools(MailCommon):
             'name': 'Alfred Astaire',
             'phone': '0456334455',
         })
-
-        cls.sources = [
-            # single email
-            'alfred.astaire@test.example.com',
-            ' alfred.astaire@test.example.com ',
-            'Fredo The Great <alfred.astaire@test.example.com>',
-            '"Fredo The Great" <alfred.astaire@test.example.com>',
-            'Fredo "The Great" <alfred.astaire@test.example.com>',
-            # multiple emails
-            'alfred.astaire@test.example.com, evelyne.gargouillis@test.example.com',
-            'Fredo The Great <alfred.astaire@test.example.com>, Evelyne The Goat <evelyne.gargouillis@test.example.com>',
-            '"Fredo The Great" <alfred.astaire@test.example.com>, evelyne.gargouillis@test.example.com',
-            '"Fredo The Great" <alfred.astaire@test.example.com>, <evelyne.gargouillis@test.example.com>',
-            # text containing email
-            'Hello alfred.astaire@test.example.com how are you ?',
-            '<p>Hello alfred.astaire@test.example.com</p>',
-            # text containing emails
-            'Hello "Fredo" <alfred.astaire@test.example.com>, evelyne.gargouillis@test.example.com',
-            'Hello "Fredo" <alfred.astaire@test.example.com> and evelyne.gargouillis@test.example.com',
-            # falsy
-            '<p>Hello Fredo</p>',
-            'j\'adore écrire des @gmail.com ou "@gmail.com" a bit randomly',
-            '',
-        ]
 
     @users('employee')
     def test_find_partner_from_emails(self):
@@ -175,88 +150,3 @@ class TestMailTools(MailCommon):
         self.user_employee.sudo().write({'email': '"Alfred Astaire" <%s>' % self.env.user.partner_id.email_normalized})
         found = self.env['res.partner']._mail_find_partner_from_emails([self.env.user.partner_id.email_formatted])
         self.assertEqual(found, [self.env.user.partner_id])
-
-    @users('employee')
-    def test_tools_email_re(self):
-        expected = [
-            # single email
-            ['alfred.astaire@test.example.com'],
-            ['alfred.astaire@test.example.com'],
-            ['alfred.astaire@test.example.com'],
-            ['alfred.astaire@test.example.com'],
-            ['alfred.astaire@test.example.com'],
-            # multiple emails
-            ['alfred.astaire@test.example.com', 'evelyne.gargouillis@test.example.com'],
-            ['alfred.astaire@test.example.com', 'evelyne.gargouillis@test.example.com'],
-            ['alfred.astaire@test.example.com', 'evelyne.gargouillis@test.example.com'],
-            ['alfred.astaire@test.example.com', 'evelyne.gargouillis@test.example.com'],
-            # text containing email
-            ['alfred.astaire@test.example.com'],
-            ['alfred.astaire@test.example.com'],
-            # text containing emails
-            ['alfred.astaire@test.example.com', 'evelyne.gargouillis@test.example.com'],
-            ['alfred.astaire@test.example.com', 'evelyne.gargouillis@test.example.com'],
-            # falsy
-            [], [], [],
-        ]
-
-        for src, exp in zip(self.sources, expected):
-            res = tools.email_re.findall(src)
-            self.assertEqual(
-                res, exp,
-                'Seems email_re is broken with %s (expected %r, received %r)' % (src, exp, res)
-            )
-
-    @users('employee')
-    def test_tools_email_split_tuples(self):
-        expected = [
-            # single email
-            [('', 'alfred.astaire@test.example.com')],
-            [('', 'alfred.astaire@test.example.com')],
-            [('Fredo The Great', 'alfred.astaire@test.example.com')],
-            [('Fredo The Great', 'alfred.astaire@test.example.com')],
-            [('Fredo The Great', 'alfred.astaire@test.example.com')],
-            # multiple emails
-            [('', 'alfred.astaire@test.example.com'), ('', 'evelyne.gargouillis@test.example.com')],
-            [('Fredo The Great', 'alfred.astaire@test.example.com'), ('Evelyne The Goat', 'evelyne.gargouillis@test.example.com')],
-            [('Fredo The Great', 'alfred.astaire@test.example.com'), ('', 'evelyne.gargouillis@test.example.com')],
-            [('Fredo The Great', 'alfred.astaire@test.example.com'), ('', 'evelyne.gargouillis@test.example.com')],
-            # text containing email -> probably not designed for that
-            [('', 'Hello alfred.astaire@test.example.comhowareyou?')],
-            [('', 'Hello alfred.astaire@test.example.com')],
-            # text containing emails -> probably not designed for that
-            [('Hello Fredo', 'alfred.astaire@test.example.com'), ('', 'evelyne.gargouillis@test.example.com')],
-            [('Hello Fredo', 'alfred.astaire@test.example.com'), ('', 'and evelyne.gargouillis@test.example.com')],
-            # falsy -> probably not designed for that
-            [], [('', "j'adore écrire des@gmail.comou"), ('', '@gmail.com')], [],
-        ]
-
-        for src, exp in zip(self.sources, expected):
-            res = tools.email_split_tuples(src)
-            self.assertEqual(
-                res, exp,
-                'Seems email_split_tuples is broken with %s (expected %r, received %r)' % (src, exp, res)
-            )
-
-    @users('employee')
-    def test_tools_single_email_re(self):
-        expected = [
-            # single email
-            ['alfred.astaire@test.example.com'],
-            [], [], [], [], # formatting issue for single email re
-            # multiple emails -> couic
-            [], [], [], [],
-            # text containing email -> couic
-            [], [],
-            # text containing emails -> couic
-            [], [],
-            # falsy
-            [], [], [],
-        ]
-
-        for src, exp in zip(self.sources, expected):
-            res = tools.single_email_re.findall(src)
-            self.assertEqual(
-                res, exp,
-                'Seems single_email_re is broken with %s (expected %r, received %r)' % (src, exp, res)
-            )
